Current Quiz saves user answers to localstorage for scoring
