+++ client/src/components/take-quiz/quiz-components/Answers.jsx
@@ -3,37 +3,38 @@ import "../styles/take-quiz.css";
 import { Button } from "@mui/material";
 
 function Answers({ page, answers }) {
-  const [ans, setAns] = useState([]);
-  useEffect(() => {
-    Object.keys(answers).forEach((key) => {
-      if (key !== "question") {
-        console.log("key", key);
-        setAns((prev) => [...prev, answers[key]]);
-      }
-    });
-  }, [answers]);
+  let keys = Object.keys(answers);
+  answers = Object.values(answers);
 
   const handleClick = (e) => {
     let selected = Object.values(e.currentTarget.firstChild)[0].stateNode.data;
-    // let key = Object.keys(e.currentTarget.firstChild)[0].stateNode.data;
-    // answer[key] = selected;
+    keys.forEach((key, index) => {
+      if (answers[index] === selected) {
+        // console.log("Selected Key", answers[index], selected, key);
+        selected = key;
+      }
+    });
     // console.log("Saving your answer for the current Question", answer);
-    localStorage.setItem(`Question ${page} Answer`, selected);
+    localStorage.setItem(`${page}`, selected);
   };
 
   return (
-    ans && (
+    answers && (
       <>
-        {ans.map((key, index) => (
-          <Button
-            key={index}
-            variant="contained"
-            color="secondary"
-            onClick={handleClick}
-          >
-            {key}
-          </Button>
-        ))}
+        {answers.map((key, index) => {
+          if (keys[index] !== "question") {
+            return (
+              <Button
+                key={index}
+                variant="contained"
+                color="secondary"
+                onClick={handleClick}
+              >
+                {answers[index]}
+              </Button>
+            );
+          }
+        })}
       </>
     )
   );
